+++ Sources/MapboxNavigationCore/Settings/RoutingConfig.swift
@@ -23,12 +23,31 @@ public struct RoutingConfig: Equatable {
     /// A time interval in which time-dependent properties of the ``RouteLeg``s of the resulting `Route`s will be
     /// refreshed.
     ///
-    /// This property is ignored unless `profileIdentifier` is `ProfileIdentifier.automobileAvoidingTraffic`. Use `nil`
-    /// value to disable the mechanism
+    /// This property is ignored unless ``profileIdentifier`` is `ProfileIdentifier.automobileAvoidingTraffic`.
+    /// Use `nil` value to disable the mechanism.
     public var routeRefreshPeriod: TimeInterval?
+
+    /// Type of routing to be used by various SDK objects when providing route calculations. Use this value to configure
+    /// online vs. offline data usage for routing.
+    ///
+    /// Default value is ``RoutingProviderSource/hybrid``
     public var routingProviderSource: RoutingProviderSource
+
+    /// Enables automatic switching to online version of the current route when possible.
+    ///
+    /// Indicates if ``NavigationController`` will attempt to detect if thr current route was build offline and if there
+    /// is an online route with the same path is available to automatically switch to it. Using online route is
+    /// beneficial due to available live data like traffic congestion, incidents, etc. Check is not performed instantly
+    /// and it is not guaranteed to receive an online version at any given period of time.
+    ///
+    /// Enabled by default.
     public var prefersOnlineRoute: Bool
 
+    @available(
+        *,
+        deprecated,
+        message: "Use 'init(alternativeRoutesDetectionConfig:fasterRouteDetectionConfig:rerouteConfig:initialManeuverAvoidanceRadius:routeRefreshPeriod:routingProviderSource:prefersOnlineRoute:)' instead."
+    )
     public init(
         alternativeRoutesDetectionSettings: AlternativeRoutesDetectionConfig? = .init(),
         fasterRouteDetectionSettings: FasterRouteDetectionConfig? = .init(),
@@ -42,6 +61,24 @@ public struct RoutingConfig: Equatable {
         self.alternativeRoutesDetectionConfig = alternativeRoutesDetectionSettings
         self.fasterRouteDetectionConfig = fasterRouteDetectionSettings
         self.rerouteConfig = rerouteSettings
+        self.initialManeuverAvoidanceRadius = initialManeuverAvoidanceRadius
+        self.routeRefreshPeriod = routeRefreshPeriod
+        self.routingProviderSource = routingProviderSource
+        self.prefersOnlineRoute = prefersOnlineRoute
+    }
+
+    public init(
+        alternativeRoutesDetectionConfig: AlternativeRoutesDetectionConfig? = .init(),
+        fasterRouteDetectionConfig: FasterRouteDetectionConfig? = .init(),
+        rerouteConfig: RerouteConfig = .init(),
+        initialManeuverAvoidanceRadius: TimeInterval = 8,
+        routeRefreshPeriod: TimeInterval? = 120,
+        routingProviderSource: RoutingProviderSource = .hybrid,
+        prefersOnlineRoute: Bool = true
+    ) {
+        self.alternativeRoutesDetectionConfig = alternativeRoutesDetectionConfig
+        self.fasterRouteDetectionConfig = fasterRouteDetectionConfig
+        self.rerouteConfig = rerouteConfig
         self.initialManeuverAvoidanceRadius = initialManeuverAvoidanceRadius
         self.routeRefreshPeriod = routeRefreshPeriod
         self.routingProviderSource = routingProviderSource